--- app/src/scroll-area/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ScrollArea';
