+++ lib/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@moondreamsdev/dreamer-ui",
-  "version": "1.7.7-test.11",
+  "version": "1.7.7-test.14",
   "publishConfig": {
     "access": "public"
   },
