+++ lib/src/components/scroll-area/index.ts
@@ -1,0 +1,2 @@
+export { default as ScrollArea } from './ScrollArea';
+export type { ScrollAreaProps } from './ScrollArea';
